@@ -3,26 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fontsource/urbanist": "^4.5.14",
-    "@testing-library/jest-dom": "^5.16.5",
+    "@fontsource/urbanist": "^4.5.17",
+    "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^26.0.24",
     "@types/node": "^15.14.9",
-    "@types/react": "^17.0.50",
-    "@types/react-dom": "^17.0.17",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
     "custom-cursor-react": "^1.0.5",
-    "framer-motion": "^7.6.1",
+    "framer-motion": "^7.10.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-particles": "^2.3.3",
+    "react-particles": "^2.12.2",
     "react-responsive-carousel": "^3.2.23",
-    "react-router-dom": "^6.4.4",
+    "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
-    "tslib": "^2.4.0",
-    "tsparticles": "^2.3.4",
-    "tsparticles-engine": "^2.3.3",
-    "typescript": "^4.8.4",
+    "tslib": "^2.6.2",
+    "tsparticles": "^2.12.0",
+    "tsparticles-engine": "^2.12.0",
+    "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -45,5 +45,17 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "eslint": "^8.54.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-standard-with-typescript": "^39.1.1",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-react": "^7.33.2",
+    "prettier": "^3.1.0"
   }
 }
