@@ -1,19 +1,23 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import reportWebVitals from './reportWebVitals';
-import { HashRouter } from "react-router-dom";
+import { HashRouter } from 'react-router-dom';
 
-ReactDOM.render(
+const rootElement = document.getElementById('root');
+
+if (rootElement !== null) {
+  const root = createRoot(rootElement);
+  root.render(
     <React.StrictMode>
-        <HashRouter>
-            <App/>
-        </HashRouter>
+      <HashRouter>
+        <App />
+      </HashRouter>
     </React.StrictMode>,
-    document.getElementById('root')
-);
+  );
+}
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
