@@ -3,24 +3,24 @@ import { motion } from 'framer-motion';
 
 import '../css/Studio.css';
 
-
 const Studio = () => {
   return (
-    <motion.div className='card'
-    initial={{x: "100%"}}
-    animate={{x: 0}}
-    transition={{
+    <motion.div
+      className="card"
+      initial={{ x: '100%' }}
+      animate={{ x: 0 }}
+      transition={{
         delay: 0.1,
         x: { duration: 1 },
-        default: { ease: "linear" }
-    }}
-    exit={{ x: "100%", opacity: 0 }}
+        default: { ease: 'linear' },
+      }}
+      exit={{ x: '100%', opacity: 0 }}
     >
-        <div className="studio-header">
+      <div className="studio-header">
         <h1>Studio</h1>
-        </div>
+      </div>
     </motion.div>
-  )
-}
+  );
+};
 
-export default Studio
+export default Studio;
