@@ -1,1 +1,1 @@
-declare module "*.pdf";
+declare module '*.pdf';
